refactor: localize runtime compat consumers
