+++ rustfs/src/server/mod.rs
@@ -24,6 +24,7 @@ mod prefix;
 mod readiness;
 mod runtime;
 mod service_state;
+mod storage_compat;
 pub mod tls_material;
 
 use tracing::warn;
