+++ rustfs/src/server/readiness.rs
@@ -12,11 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use super::storage_compat::{
+    Endpoint, EndpointServerPools, get_global_endpoints_opt, get_global_lock_clients, is_dist_erasure,
+    resolve_object_store_handle,
+};
 use crate::server::{ServiceState, ServiceStateManager};
 use crate::server::{has_path_prefix, is_table_catalog_path};
-use crate::storage_compat::is_dist_erasure;
-use crate::storage_compat::resolve_object_store_handle;
-use crate::storage_compat::{get_global_endpoints_opt, get_global_lock_clients};
 use bytes::Bytes;
 use http::{Request as HttpRequest, Response, StatusCode};
 use http_body::Body;
@@ -496,10 +497,10 @@ async fn collect_storage_readiness_uncached() -> bool {
     }
 }
 
-fn set_lock_quorum_status(online_hosts: &HashSet<String>, set_endpoints: &[crate::storage_compat::Endpoint]) -> LockQuorumStatus {
+fn set_lock_quorum_status(online_hosts: &HashSet<String>, set_endpoints: &[Endpoint]) -> LockQuorumStatus {
     let total_clients = set_endpoints
         .iter()
-        .map(crate::storage_compat::Endpoint::host_port)
+        .map(Endpoint::host_port)
         .filter(|host| !host.is_empty())
         .collect::<HashSet<_>>();
     let total_clients_len = total_clients.len();
@@ -522,10 +523,7 @@ fn set_lock_quorum_status(online_hosts: &HashSet<String>, set_endpoints: &[crate
     }
 }
 
-fn aggregate_lock_quorum_status(
-    pool_endpoints: &crate::storage_compat::EndpointServerPools,
-    online_hosts: &HashSet<String>,
-) -> LockQuorumStatus {
+fn aggregate_lock_quorum_status(pool_endpoints: &EndpointServerPools, online_hosts: &HashSet<String>) -> LockQuorumStatus {
     let mut connected_clients = 0usize;
     let mut total_clients = 0usize;
     let mut required_quorum = 0usize;
@@ -839,8 +837,7 @@ mod tests {
 
     #[test]
     fn aggregate_lock_quorum_status_requires_each_set_to_meet_quorum() {
-        use crate::storage_compat::Endpoint;
-        use crate::storage_compat::{EndpointServerPools, Endpoints, PoolEndpoints};
+        use crate::server::storage_compat::{Endpoint, EndpointServerPools, Endpoints, PoolEndpoints};
 
         let endpoints = vec![
             Endpoint {
@@ -897,8 +894,7 @@ mod tests {
 
     #[test]
     fn aggregate_lock_quorum_status_fails_when_any_set_loses_quorum() {
-        use crate::storage_compat::Endpoint;
-        use crate::storage_compat::{EndpointServerPools, Endpoints, PoolEndpoints};
+        use crate::server::storage_compat::{Endpoint, EndpointServerPools, Endpoints, PoolEndpoints};
 
         let endpoints = vec![
             Endpoint {
